requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
accelerate=0.24.1=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
appnope=0.1.3=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
bardapi=0.1.38=pypi_0
beautifulsoup4=4.12.2=pypi_0
black=23.11.0=pypi_0
blinker=1.7.0=pypi_0
browser-cookie3=0.19.1=pypi_0
bzip2=1.0.8=h93a5062_5
ca-certificates=2023.7.22=hf0a4a13_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.0=pypi_0
dataclasses-json=0.6.2=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
deep-translator=1.11.4=pypi_0
distro=1.8.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.1=pypi_0
filelock=3.13.1=pyhd8ed1ab_0
flask=3.0.0=pypi_0
flask-cors=4.0.0=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
gmp=6.3.0=h965bd2d_0
gmpy2=2.1.2=py39h0b4f9c6_1
google-api-core=2.14.0=pypi_0
google-auth=2.23.4=pypi_0
google-cloud-core=2.3.3=pypi_0
google-cloud-translate=3.12.1=pypi_0
googleapis-common-protos=1.61.0=pypi_0
grpcio=1.59.2=pypi_0
grpcio-status=1.59.2=pypi_0
h11=0.14.0=pypi_0
h2=4.1.0=pypi_0
hpack=4.0.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.25.1=pypi_0
huggingface-hub=0.17.3=pypi_0
hyperframe=6.0.1=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
ipykernel=6.26.0=pypi_0
ipython=8.17.2=pypi_0
ipywidgets=8.1.1=pypi_0
isort=5.12.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.5.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
langchain=0.0.334=pypi_0
langdetect=1.0.9=pypi_0
langsmith=0.0.63=pypi_0
libcxx=16.0.6=h4653b0c_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.44.0=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
loguru=0.7.2=pypi_0
lz4=4.3.2=pypi_0
markupsafe=2.1.3=py39h0f82c59_1
marshmallow=3.20.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mpc=1.3.1=h91ba8db_0
mpfr=4.2.1=h9546428_0
mpmath=1.3.0=pyhd8ed1ab_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h463b476_2
nest-asyncio=1.5.8=pypi_0
networkx=3.2.1=pyhd8ed1ab_0
numpy=1.26.1=pypi_0
openai=1.2.3=pypi_0
openssl=3.1.4=h0d3ecfb_0
packaging=23.2=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
peft=0.6.1=pypi_0
pexpect=4.8.0=pypi_0
pip=23.3.1=pyhd8ed1ab_0
platformdirs=4.0.0=pypi_0
prompt-toolkit=3.0.40=pypi_0
proto-plus=1.22.3=pypi_0
protobuf=4.25.0=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycryptodomex=3.19.0=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pygments=2.16.1=pypi_0
python=3.9.18=hfa1ae8a_0_cpython
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python_abi=3.9=4_cp39
pytorch=2.1.0=py3.9_0
pyyaml=6.0.1=py39h0f82c59_1
pyzmq=25.1.1=pypi_0
readline=8.2=h92ec313_1
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
rsa=4.9=pypi_0
ruff=0.1.5=pypi_0
safetensors=0.4.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.23=pypi_0
stack-data=0.6.3=pypi_0
sympy=1.12=pypyh9d50eac_103
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.13=h5083fa2_1
tokenizers=0.14.1=pypi_0
tomli=2.0.1=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pypi_0
transformers=4.35.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=2.0.7=pypi_0
wcwidth=0.2.9=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.3=pyhd8ed1ab_0
widgetsnbextension=4.0.9=pypi_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
